requirements.txt
langchain[all]==0.1.13
langchain-openai==0.1.7
langchain-huggingface==0.0.3
langchain-chroma==0.1.4
chromadb==0.5.23
sentence-transformers==4.1.0
openai==1.78.1
tiktoken==0.9.0
python-dotenv==1.1.0
pypdf==5.5.0
pdfminer.six==20221105
pi-heif==0.22.0
unstructured==0.17.2
unstructured-inference==0.7.26
requests>=2.32.0
pdf2image==1.17.0
unstructured-pytesseract==0.3.15
